vector_search/requirements.txt
# Vector Database
chromadb>=0.5.6

# Google Gemini Embeddings
google-generativeai>=0.8.0

# Document Processing & Chunking
langchain-text-splitters>=0.3.0
tiktoken>=0.7.0  # Token counting

# FastAPI for REST API
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=24.0.0
